==> 1_data_processing/paralog_features/04_process_gtex_expression.ipynb <==
## Process GTEx RNA-Seq data

Data from https://www.gtexportal.org/home/

NOTE: Large files, takes some time (~20-30 mins) to run.

import os
import pandas as pd
import re
import numpy as np
import scipy.stats as stats

get_data_path = lambda folders, fname: os.path.normpath(os.environ['3RD_PARTY_DIR']+'/'+'/'.join(folders) +'/'+ fname)
get_local_data_path = lambda folders, fname: os.path.normpath('../../local_data/' +'/'.join(folders) +'/'+ fname)

# Can read the zipped file directly with pandas
file_gtex_expr = get_data_path(['GTEx'], 'GTEx_Analysis_2017-06-05_v8_RNASeQCv1.1.9_gene_tpm.gct.gz')
file_paralog_pairs = get_local_data_path(['processed', 'ensembl93'], 'all_pairs.csv')

# Output
file_gtex_expr_features = get_local_data_path(['processed', 'paralog_features'], 'gtex_expr_features.csv')

### Extract expression for genes that are in a paralog pair
Only needs to run once, filtered data is saved

all_pairs = pd.read_csv(file_paralog_pairs)[['A1_ensembl', 'A2_ensembl']]
print('N pairs:', all_pairs.shape[0])
paralog_genes = all_pairs.A1_ensembl.unique()
print('N genes:', len(paralog_genes))
all_pairs[:1]

# Skip first 2 rows of metadata in gct file
# Read X rows at a time and only keep expression data for paralogs
df = pd.DataFrame()
for chunk in pd.read_csv(file_gtex_expr, sep='\t', skiprows=2, iterator=True, chunksize=1000):
    chunk_df = chunk.assign(ensembl_id = chunk.Name.apply(lambda x: x.split('.')[0])).drop(columns=['Description'])
    df = pd.concat([df, chunk_df[chunk_df.ensembl_id.isin(paralog_genes)].set_index('ensembl_id')])

df = df.drop(columns=['Name'])

### Clean up expression data

print('N genes:', df.shape[0])
print('N samples:', df.shape[1])
df[:1]

# Some genes are duplicated but it seems one of them is always 0 so just keep the max value
# The same gene on X and Y chromosome?
duplicate_indices = df.index[df.index.duplicated()]
display(duplicate_indices)

df.loc['ENSG00000182378']

# Drop duplicate rows where expression is always 0
expr_data = df[~((df == 0).all(axis=1) & df.index.isin(duplicate_indices))]

%reset_selective -f "df"

print('N genes:', expr_data.shape[0])
expr_data[:1]

### Calculate expression features of paralog pairs
Min mean, max mean, correlation

mean_expr = pd.DataFrame(expr_data.mean(axis=1), columns=['mean_expr']).reset_index()
mean_expr[:1]

# Merge mean
pairs_expr = pd.merge(all_pairs, mean_expr.rename(columns={'ensembl_id':'A1_ensembl', 'mean_expr':'A1_mean_expr'}), how='left')
pairs_expr = pd.merge(pairs_expr, mean_expr.rename(columns={'ensembl_id':'A2_ensembl', 'mean_expr':'A2_mean_expr'}), how='left')

print('Pairs where both genes have expr:', 
      pairs_expr[(~pairs_expr.A1_mean_expr.isna()) & (~pairs_expr.A2_mean_expr.isna())].shape[0], '/', pairs_expr.shape[0])
pairs_expr[:1]

# Filter pairs down to those where both have expression (not NA and not 0)
pairs_expr_subset = pairs_expr[(~pairs_expr.A1_mean_expr.isna()) & (~pairs_expr.A2_mean_expr.isna()) & 
                               (pairs_expr.A1_mean_expr!=0) & (pairs_expr.A2_mean_expr!=0)]
print('N:', pairs_expr_subset.shape[0])

def compute_spearman_corr(pair):
    return stats.spearmanr(expr_data.loc[pair.A1_ensembl], expr_data.loc[pair.A2_ensembl])[0]

features = pairs_expr_subset.assign(spearman_corr = pairs_expr_subset.apply(compute_spearman_corr, axis=1))
features[:1]

def compute_pearson_corr(pair):
    return stats.pearsonr(expr_data.loc[pair.A1_ensembl], expr_data.loc[pair.A2_ensembl])[0]

# NaNs in results are due to one of the genes in the pair having all 0 expression - PearsonRConstantInputWarning
features2 = features.assign(pearson_corr = features.apply(compute_pearson_corr, axis=1))
features2[:1]

# Put all expr features together
all_expr_features = pd.merge(pairs_expr, features2[['A1_ensembl','A2_ensembl','spearman_corr','pearson_corr']], how='left')
assert(all_expr_features.shape[0] == pairs_expr.shape[0])
all_expr_features[:1]

all_expr_features.to_csv(file_gtex_expr_features, index=0)
